Feature together

+++ hello_world/lib/models/foodDiary.dart
@@ -3,12 +3,11 @@ class FoodDiary {
   String foodDiaryId;
   String userId;
   DateTime foodDiaryDate;
-  int savedCalorieGoal;
-  int savedFatGoal;
-  int savedCarbGoal;
-  int savedProteinGoal;
+  double savedCalorieGoal;
+  double savedFatGoal;
+  double savedCarbGoal;
+  double savedProteinGoal;
   int caloriesBurned;
-  //final List<String> meals;
 
   FoodDiary({ this.foodDiaryId, this.userId, this.foodDiaryDate, this.savedCalorieGoal, this.savedFatGoal, this.savedCarbGoal, this.savedProteinGoal, this.caloriesBurned});
 }
